Corrections pmssm_lem
 - Chargino interactions with photon simplified
 - theta_W expressions replaced
 - file name conventions changed
