+++ src/marty/models/CMakeLists.txt
@@ -12,19 +12,19 @@ set(ALL_SOURCES
     nmfv.h nmfv.cpp
 
     pmssm_lem.h
-    pmssm_lem/pmssm_content.cpp
     pmssm_lem/pmssm_data.h
-    pmssm_lem/pmssm_interaction.cpp
-    pmssm_lem/pmssm_kinetic.cpp
     pmssm_lem/pmssm_lem.h
-    pmssm_lem/pmssm_mass.cpp
+    pmssm_lem/pmssm_lem.content.cpp
+    pmssm_lem/pmssm_lem.interaction.cpp
+    pmssm_lem/pmssm_lem.kinetic.cpp
+    pmssm_lem/pmssm_lem.mass.cpp
 )
 
 SET_SOURCE_FILES_PROPERTIES(
-    pmssm_lem/pmssm_content.cpp
-    pmssm_lem/pmssm_interaction.cpp
-    pmssm_lem/pmssm_kinetic.cpp
-    pmssm_lem/pmssm_mass.cpp
+    pmssm_lem/pmssm_lem.content.cpp
+    pmssm_lem/pmssm_lem.interaction.cpp
+    pmssm_lem/pmssm_lem.kinetic.cpp
+    pmssm_lem/pmssm_lem.mass.cpp
     PROPERTIES COMPILE_FLAGS -O0)
 
 add_library(marty-models OBJECT ${ALL_SOURCES})
